// src/app/Models/student-trainings.ts
export class StudentTrainings {
    trainerCourseId: Number;
    userId: Number;
    ratings: Number;
    courseName: string;
    batchName: string;
    mentorName: string;
    mentorFee: number;
    studentFee: number;
    courseStatus: string;
    skillName: string;
    courseId: number;
    userName: string;
}